06/04

+++ libft/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: evlad <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/04 13:49:20 by evlad             #+#    #+#             */
-/*   Updated: 2016/11/11 15:25:20 by evlad            ###   ########.fr       */
+/*   Updated: 2017/04/06 20:15:51 by evlad            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+intmax_t	ft_atoi(const char *str)
 {
 	int i;
-	int number;
+	intmax_t number;
 	int sign;
 
 	i = 0;
